requirements.txt
python-telegram-bot==20.7
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
pytz==2023.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0